requirements/base.txt
amqp==5.0.6
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cfgv==3.3.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.3
Django==3.2.9
django-environ==0.8.1
django-filter==21.1
django-redis==5.1.0
django-threadlocals==0.10
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.10.1
drf-yasg==1.20.0
facebook-sdk==3.1.0
filelock==3.4.0
identify==2.4.0
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.2
Markdown==3.3.6
MarkupSafe==2.0.1
nodeenv==1.6.0
packaging==21.3
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
psycopg2==2.9.2
PyJWT==2.3.0
pyparsing==3.0.6
python-http-client==3.3.3
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sendgrid==6.9.1
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==2.64.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
zipp==3.6.0
